Fix broken tests, only do remote things when asked

+++ configure_hosts.sh
@@ -89,84 +89,93 @@ echo "*** Configuration file is okay."
 
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~ CREATE BATCH FILE ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-echo "*** Creating batch file..."
-BATCH_FILE=`mktemp`
+REMOTE="${APTITUDE_PACKAGES}${REMOTE_DIRECTORIES}${REMOTE_CONFIGURATIONS}${REMOTE_COMMANDS}"
+if [[ ! -z $REMOTE ]]
+then
+    echo "*** Creating batch file..."
+    BATCH_FILE=`mktemp`
 
 
-# Change working directory
-echo "echo \"--> cd $WORKING_DIRECTORY\"" > $BATCH_FILE
-echo "cd $WORKING_DIRECTORY" >> $BATCH_FILE
+    # Change working directory
+    echo "echo \"--> cd $WORKING_DIRECTORY\"" > $BATCH_FILE
+    echo "cd $WORKING_DIRECTORY" >> $BATCH_FILE
 
-# Install some packages
-if [[ ! -z $APTITUDE_PACKAGES ]]; then
-    echo "echo \"--> apt-get update\"" >> $BATCH_FILE
-    echo "apt-get update" >> $BATCH_FILE
-    echo "echo \"--> apt-get install -y $APTITUDE_PACKAGES\"" >> $BATCH_FILE
-    echo "apt-get install -y $APTITUDE_PACKAGES" >> $BATCH_FILE
-    echo 'if [ $? -ne 0 ]; then echo "Failed to install packages, terminating"; exit 1; fi' >> $BATCH_FILE
-fi
+    # Install some packages
+    if [[ ! -z $APTITUDE_PACKAGES ]]; then
+        echo "echo \"--> apt-get update\"" >> $BATCH_FILE
+        echo "apt-get update" >> $BATCH_FILE
+        echo "echo \"--> apt-get install -y $APTITUDE_PACKAGES\"" >> $BATCH_FILE
+        echo "apt-get install -y $APTITUDE_PACKAGES" >> $BATCH_FILE
+        echo 'if [ $? -ne 0 ]; then echo "Failed to install packages, terminating"; exit 1; fi' >> $BATCH_FILE
+    fi
 
-# Create some directories
-if [[ ! -z "${REMOTE_DIRECTORIES[@]}" ]]; then
-    for dir in "${REMOTE_DIRECTORIES[@]}"; do
-        echo "echo \"--> mkdir -p $dir\"" >> $BATCH_FILE
-        echo "mkdir -p $dir" >> $BATCH_FILE
+    # Create some directories
+    if [[ ! -z "${REMOTE_DIRECTORIES[@]}" ]]; then
+        for dir in "${REMOTE_DIRECTORIES[@]}"; do
+            echo "echo \"--> mkdir -p $dir\"" >> $BATCH_FILE
+            echo "mkdir -p $dir" >> $BATCH_FILE
+        done
+    fi
+
+    CONFIG_FILES=()
+    # Install some config files
+    if [[ ! -z "${REMOTE_CONFIGURATIONS[@]}" ]]; then
+        for pair in "${REMOTE_CONFIGURATIONS[@]}"; do
+            filename=${pair% *}
+            dir=${pair#* }
+            echo "echo \"--> mv $filename $dir\"" >> $BATCH_FILE
+            echo "mv $filename $dir" >> $BATCH_FILE
+
+            CONFIG_FILES+=($filename)  # Keep track of our config files
+        done
+    fi
+
+    # Run remote commands. Since these are user-provided, they might fail, so
+    # immediately exit if that happens.
+    for comm in "${REMOTE_COMMANDS[@]}"; do
+        echo "echo \"--> $comm\"" >> $BATCH_FILE
+        echo "$comm" >> $BATCH_FILE
+        echo 'if [ $? -ne 0 ]; then echo "Previous command failed, terminating"; exit 1; fi' >> $BATCH_FILE
     done
+
+    # Clean up after ourselves; we'll always name this file config.sh
+    echo "rm $WORKING_DIRECTORY/config.sh" >> $BATCH_FILE
+
+    echo "*** Batch file created"
+else
+    echo "*** No need to log in to remote machine..."
 fi
-
-CONFIG_FILES=()
-# Install some config files
-if [[ ! -z "${REMOTE_CONFIGURATIONS[@]}" ]]; then
-    for pair in "${REMOTE_CONFIGURATIONS[@]}"; do
-        filename=${pair% *}
-        dir=${pair#* }
-        echo "echo \"--> mv $filename $dir\"" >> $BATCH_FILE
-        echo "mv $filename $dir" >> $BATCH_FILE
-    
-        CONFIG_FILES+=($filename)  # Keep track of our config files
-    done
-fi
-
-# Run remote commands. Since these are user-provided, they might fail, so 
-# immediately exit if that happens.
-for comm in "${REMOTE_COMMANDS[@]}"; do
-    echo "echo \"--> $comm\"" >> $BATCH_FILE
-    echo "$comm" >> $BATCH_FILE
-    echo 'if [ $? -ne 0 ]; then echo "Previous command failed, terminating"; exit 1; fi' >> $BATCH_FILE
-done
-
-# Clean up after ourselves; we'll always name this file config.sh
-echo "rm $WORKING_DIRECTORY/config.sh" >> $BATCH_FILE
-
-echo "*** Batch file created"
 
 
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~ EXECUTE COMMANDS ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 for host in $HOSTS; do
-    echo "*** Configuring host $host ***"
-    echo "*** Copying over ssh key..."
-    ssh-copy-id -i $ID root@$host
+    export host
+    if [[ ! -z $REMOTE ]]; then
+        echo "*** Configuring host $host ***"
+        echo "*** Copying over ssh key..."
+        ssh-copy-id -i $ID root@$host
 
-    echo "*** Copying over configuration files..."
-    ssh root@$host "mkdir -p $WORKING_DIRECTORY"
-    scp $BATCH_FILE root@$host:$WORKING_DIRECTORY/config.sh
-    die_on_failure "Failed to copy over batch file. Exiting."
+        echo "*** Copying over configuration files..."
+        ssh root@$host "mkdir -p $WORKING_DIRECTORY"
+        scp $BATCH_FILE root@$host:$WORKING_DIRECTORY/config.sh
+        die_on_failure "Failed to copy over batch file. Exiting."
 
-    if [[ ! -z "${CONFIG_FILES[@]}" ]]; then
-        scp ${CONFIG_FILES[@]} root@$host:$WORKING_DIRECTORY
-        die_on_failure "Failed to copy over configuration files. Exiting."
+        if [[ ! -z "${CONFIG_FILES[@]}" ]]; then
+            scp ${CONFIG_FILES[@]} root@$host:$WORKING_DIRECTORY
+            die_on_failure "Failed to copy over configuration files. Exiting."
+        fi
+
+        echo -e "\n*** Applying configuration to remote machine..."
+        ssh root@$host "bash $WORKING_DIRECTORY/config.sh"
+        die_on_failure "Remote execution failed. Exiting."
     fi
-
-    echo -e "\n*** Applying configuration to remote machine..."
-    ssh root@$host "bash $WORKING_DIRECTORY/config.sh"
-    die_on_failure "Remote execution failed. Exiting."
 
     if [ ! -z "${TEST_COMMANDS[@]}" ]; then
         echo "*** Running test commands..."
         for comm in "${TEST_COMMANDS[@]}"; do
             echo $comm
-            if [[ $comm ]]; then
+            if `bash -c "$comm"`; then
                 echo "*** Test succeeded."
             else
                 echo "*** Test failed." 1>&2
@@ -175,4 +184,6 @@ for host in $HOSTS; do
     fi
 done
 
-rm $BATCH_FILE
+if [[ ! -z $REMOTE ]]; then
+    rm $BATCH_FILE
+fi
